Provide 'copied' feedback via controlled tooltip
